src/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
python-dotenv==1.0.0
crewai==1.3.0
crewai[tools]==1.3.0
pydantic==2.5.0
